move 31x and db base shims
